+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [3.5.1]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [3.5.2]
 This version allows Globus to be configured without a local endpoint.
 
 ### Fixed